Delete killdd.py
