--- userp/killdd.py
@@ -1,7 +1,0 @@
-import re
-
-
-def ddk(str1):
-    a = re.sub('<dd>', '', str1)
-    a = re.sub('</dd>', '', a)
-    return (a)
